# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: AI will API
  description: |
    AI will - AI音声会話アプリのREST API仕様書

    ## 認証
    JWT Bearer Tokenを使用します。
    ```
    Authorization: Bearer {access_token}
    ```

    ## エラーレスポンス
    すべてのエラーは統一フォーマットで返却されます。
    ```json
    {
      "error": {
        "code": "error_code",
        "message": "エラーメッセージ",
        "details": []
      }
    }
    ```

    ## ページング
    Cursor方式を採用。`cursor`と`limit`パラメータを使用。
    ```json
    {
      "data": [...],
      "pagination": {
        "next_cursor": "eyJpZCI6...",
        "has_more": true
      }
    }
    ```

    ## 冪等性（Idempotency）
    購入系・ジョブ作成系APIでは`Idempotency-Key`ヘッダーが必須です。
    - キー: クライアント生成のUUID
    - 有効期限: 24時間
    - 同一キーで再リクエスト時は同一レスポンスを返却

    ## ストリーミング（SSE）
    会話APIでは Server-Sent Events によるストリーミングをサポートします。
    - エンドポイント: `POST /threads/{id}/messages:stream`
    - イベント: `message_start`, `content_delta`, `message_done`, `error`

    ## 主要なユースケース例
    - **Example 1**: 会話メッセージ送信（`POST /threads/{id}/messages`）
    - **Example 2**: 購入反映・Idempotency-Key付き（`POST /purchases`）
    - **Example 3**: クリップ作成（`POST /clips`）
  version: 1.3.0
  contact:
    name: AI will Support
    email: support@aiwill.app
  license:
    name: Proprietary
    url: https://aiwill.app/terms
servers:
  - url: https://api.aiwill.app/v1
    description: Production
  - url: https://api.staging.aiwill.app/v1
    description: Staging
...
paths:
...
  /purchases:
    post:
      tags: [Purchase]
      summary: 購入
      description: |
        Packを購入します。

        ## 冪等性

        `Idempotency-Key` ヘッダーが**必須**です。

        ```http
        POST /v1/purchases
        Authorization: Bearer {access_token}
        Idempotency-Key: 550e8400-e29b-41d4-a716-446655440000
        Content-Type: application/json
        ```

        ### 冪等性キーの動作
        - 同一キーで24時間以内に再リクエストした場合、同一レスポンスを返却
        - キー未指定時は `400 Bad Request` (`idempotency_key_required`)
        - 処理中に再リクエストした場合は `409 Conflict` (`request_in_progress`)

        ## 決済フロー
        1. クライアントが決済プロバイダー（Stripe/Apple/Google）で決済を完了
        2. 決済プロバイダーからレシートを取得
        3. 本APIにレシートを送信してサーバー側で検証
        4. 検証成功で利用権（Entitlement）を付与
      operationId: createPurchase
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/idempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePurchaseRequest'
            examples:
              stripe_purchase:
                summary: Stripe決済での購入
                description: |
                  **Example 2: 購入反映（Idempotency-Key付き）**

                  Stripe決済で Pack を購入する例。
                  Idempotency-Key ヘッダーにより、ネットワーク障害時の再送でも
                  重複課金を防止します。
                value:
                  pack_id: pack_550e8400-e29b-41d4-a716-446655440000
                  payment_method: stripe
                  receipt_data: pi_3NkVz2Abc123XYZ_secret_abc123xyz
              apple_purchase:
                summary: Apple In-App Purchase
                value:
                  pack_id: pack_550e8400-e29b-41d4-a716-446655440000
                  payment_method: apple
                  receipt_data: MIITtgYJKoZIhvc...（Base64エンコードされたレシート）
              google_purchase:
                summary: Google Play Billing
                value:
                  pack_id: pack_550e8400-e29b-41d4-a716-446655440000
                  payment_method: google
                  receipt_data: '{"orderId":"GPA.1234-5678-9012","packageName":"app.aiwill","productId":"pack_001"}'
      responses:
        '201':
          description: 購入成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PurchaseResponse'
              examples:
                purchase_completed:
                  summary: 購入完了（Stripe）
                  description: |
                    **Example 2: 購入成功レスポンス**

                    購入が正常に完了し、利用権（Entitlement）が付与された状態。
                    以降、このPackのキャラクターと会話が可能になります。
                  value:
                    purchase:
                      id: purchase_550e8400-e29b-41d4-a716-446655440000
                      pack_id: pack_550e8400-e29b-41d4-a716-446655440000
                      amount: 500
                      currency: JPY
                      payment_method: stripe
                      status: completed
                      created_at: '2026-01-31T12:00:00Z'
                    entitlement:
                      id: ent_550e8400-e29b-41d4-a716-446655440000
                      entitlement_type: pack
                      entitlement_id: pack_550e8400-e29b-41d4-a716-446655440000
                      granted_at: '2026-01-31T12:00:00Z'
                      revoked_at: null
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '409':
          description: 購入済みまたは処理中
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                already_purchased:
                  summary: 購入済み
                  value:
                    error:
                      code: already_purchased
                      message: このPackは既に購入済みです
                      details: []
                request_in_progress:
                  summary: 処理中（冪等性キー重複）
                  value:
                    error:
                      code: request_in_progress
                      message: 同一の冪等性キーでリクエストが処理中です
                      details: []
  /purchases:restore:
    post:
      tags: [Purchase]
      summary: 購入復元
      description: |
        ストアの購入状態を復元します。

        **冪等性:** `Idempotency-Key`ヘッダー必須
      operationId: restorePurchases
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/idempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RestorePurchaseRequest'
      responses:
        '200':
          description: 復元成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestorePurchaseResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /entitlements:
    get:
      tags: [Purchase]
      summary: 利用権一覧取得
      description: ユーザーが保持する利用権の一覧を取得します。
      operationId: listEntitlements
      security:
        - BearerAuth: []
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EntitlementListResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

  # ----------------------------------------------------------
  # Memory (Clips)
  # ----------------------------------------------------------
...
  /report-reasons:
    get:
      tags: [Safety]
      summary: 通報理由一覧取得
      description: 通報時に選択できる理由の一覧を取得します。
      operationId: listReportReasons
      security:
        - BearerAuth: []
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportReasonListResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /reports:
    post:
      tags: [Safety]
      summary: 通報
      description: コンテンツやユーザーを通報します。
      operationId: createReport
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReportRequest'
      responses:
        '201':
          description: 通報受付
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          description: 重複通報
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /blocks:
    get:
      tags: [Safety]
      summary: ブロック一覧取得
      description: ユーザーがブロックしている対象の一覧を取得します。
      operationId: listBlocks
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/cursor'
        - $ref: '#/components/parameters/limit'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BlockListResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    post:
      tags: [Safety]
      summary: ブロック
      description: コンテンツやクリエイターをブロックします。
      operationId: createBlock
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBlockRequest'
      responses:
        '201':
          description: ブロック成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BlockResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '409':
          $ref: '#/components/responses/ConflictError'
  /blocks/{block_id}:
    delete:
      tags: [Safety]
      summary: ブロック解除
      description: 指定したブロックを解除します。
      operationId: deleteBlock
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/blockId'
      responses:
        '204':
          description: 解除成功
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'

  # ----------------------------------------------------------
  # Privacy
  # ----------------------------------------------------------
...
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Bearer Token

  # ----------------------------------------------------------
  # Parameters
  # ----------------------------------------------------------
  parameters:
...
    blockId:
      name: block_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Block ID
...
    cursor:
      name: cursor
      in: query
      schema:
        type: string
      description: ページングカーソル
    limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: 取得件数
...
    idempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        format: uuid
      description: 冪等性キー（UUID）

  # ----------------------------------------------------------
  # Responses
  # ----------------------------------------------------------
  responses:
    ValidationError:
      description: バリデーションエラー
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: validation_error
              message: リクエストの検証に失敗しました
              details:
                - field: email
                  code: invalid_format
                  message: 有効なメールアドレス形式で入力してください
    UnauthorizedError:
      description: 認証エラー
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: unauthenticated
              message: 認証が必要です
              details: []
...
    NotFoundError:
      description: リソースが見つからない
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: not_found
              message: リソースが見つかりません
              details: []
    ConflictError:
      description: 競合エラー
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: conflict
              message: リソースが既に存在します
              details: []
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: エラーコード
            message:
              type: string
              description: エラーメッセージ
            details:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  code:
                    type: string
                  message:
                    type: string
    Pagination:
      type: object
      properties:
        next_cursor:
          type: string
          nullable: true
          description: 次ページのカーソル
        has_more:
          type: boolean
          description: 次ページの有無

    # Auth
...
    CreatePurchaseRequest:
      type: object
      required: [pack_id, payment_method]
      properties:
        pack_id:
          type: string
          format: uuid
        payment_method:
          type: string
          enum: [stripe, apple, google]
        receipt_data:
          type: string
          description: 決済プロバイダーからのレシート
    Purchase:
      type: object
      description: 購入情報
      properties:
        id:
          type: string
          format: uuid
          description: 購入ID
        pack_id:
          type: string
          format: uuid
          description: 購入対象のPack ID
        amount:
          type: integer
          description: 購入金額
          example: 500
        currency:
          type: string
          description: 通貨コード（ISO 4217）
          enum: [JPY, USD]
          default: JPY
          example: JPY
        payment_method:
          type: string
          description: 決済方法
          enum: [stripe, apple, google]
          example: stripe
        status:
          type: string
          enum: [pending, completed, failed, refunded]
          description: |
            購入ステータス:
            - `pending`: 処理中
            - `completed`: 完了
            - `failed`: 失敗
            - `refunded`: 返金済み
        created_at:
          type: string
          format: date-time
    Entitlement:
      type: object
      properties:
        id:
          type: string
          format: uuid
        entitlement_type:
          type: string
          enum: [pack]
        entitlement_id:
          type: string
          format: uuid
        granted_at:
          type: string
          format: date-time
        revoked_at:
          type: string
          format: date-time
          nullable: true
    PurchaseResponse:
      type: object
      properties:
        purchase:
          $ref: '#/components/schemas/Purchase'
        entitlement:
          $ref: '#/components/schemas/Entitlement'
    RestorePurchaseRequest:
      type: object
      required: [payment_method]
      properties:
        payment_method:
          type: string
          enum: [stripe, apple, google]
        receipt_data:
          type: string
    RestorePurchaseResponse:
      type: object
      properties:
        restored_count:
          type: integer
        entitlements:
          type: array
          items:
            $ref: '#/components/schemas/Entitlement'
    EntitlementListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Entitlement'

    # Memory (Clips)
...
    ReportReason:
      type: object
      properties:
        id:
          type: string
          format: uuid
        code:
          type: string
        name:
          type: string
    ReportReasonListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/ReportReason'
    CreateReportRequest:
      type: object
      required: [target_type, target_id, reason_id]
      properties:
        target_type:
          type: string
          enum: [character, message, pack, creator]
        target_id:
          type: string
          format: uuid
        reason_id:
          type: string
          format: uuid
        comment:
          type: string
          maxLength: 500
    Report:
      type: object
      properties:
        id:
          type: string
          format: uuid
        target_type:
          type: string
        target_id:
          type: string
          format: uuid
        reason:
          $ref: '#/components/schemas/ReportReason'
        status:
          type: string
          enum: [open, in_progress, resolved, rejected]
        created_at:
          type: string
          format: date-time
    ReportResponse:
      type: object
      properties:
        report:
          $ref: '#/components/schemas/Report'
        message:
          type: string
    CreateBlockRequest:
      type: object
      required: [target_type, target_id]
      properties:
        target_type:
          type: string
          enum: [user, creator, character, tag]
        target_id:
          type: string
          format: uuid
    Block:
      type: object
      properties:
        id:
          type: string
          format: uuid
        target_type:
          type: string
        target_id:
          type: string
          format: uuid
        target_name:
          type: string
        created_at:
          type: string
          format: date-time
    BlockResponse:
      type: object
      properties:
        block:
          $ref: '#/components/schemas/Block'
    BlockListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Block'
        pagination:
          $ref: '#/components/schemas/Pagination'

    # Privacy - Jobs
...